closes #5
I've kept the same version of Salmon. This kind of problem is fairly
common with Conda packages so I don't see that glossing over it by
picking a better behaved package would be helpful. Also as new packages
are added to Bioconda this may break further, but that's outside our
control.